pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "surfacescope"
version = "0.1.0"
description = "Authorized attack surface inventory and exposure prioritization toolkit"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  {name = "Amadu M Balde", email = "amadubalde.cyber@gmail.com"}
]
license = "MIT"
dependencies = [
  "requests>=2.32.0",
  "dnspython>=2.6.1",
  "jinja2>=3.1.4",
  "packaging>=24.0",
  "click>=8.1.7",
  "rich>=13.7.1",
  "pytest>=8.2.0",
  "cryptography>=42.0.0",
  "beautifulsoup4>=4.12.3"
]

[project.scripts]
surfacescope = "surfacescope.cli:main"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.setuptools]
packages = ["surfacescope", "surfacescope.modules"]
